\name{threePG-package}
\alias{threePG-package}
\alias{threePG}
\docType{package}
\title{
\packageTitle{threePG}
}
\description{
\packageDescription{threePG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{threePG}
\packageIndices{threePG}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{threePG}

Maintainer: \packageMaintainer{threePG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
